refactor(recommendation): 추천 리스트 페이지 책임 분리 (#268)
